Updated version, based on Andriy's comment.
