+++ src/EFCore.Cosmos/Infrastructure/Internal/CosmosModelRuntimeInitializerDependencies.cs
@@ -1,9 +1,9 @@
-// Licensed to the .NET Foundation under one or more agreements.
+﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using Microsoft.EntityFrameworkCore.Cosmos.Metadata.Internal;
 
-namespace Microsoft.EntityFrameworkCore.Cosmos.ValueGeneration.Internal;
+namespace Microsoft.EntityFrameworkCore.Infrastructure.Internal;
 
 /// <summary>
 ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -11,7 +11,7 @@ namespace Microsoft.EntityFrameworkCore.Cosmos.ValueGeneration.Internal;
 ///     any release. You should only use it directly in your code with extreme caution and knowing that
 ///     doing so can result in application failures when updating to a new Entity Framework Core release.
 /// </summary>
-public class IdValueGenerator : ValueGenerator
+public sealed record CosmosModelRuntimeInitializerDependencies
 {
     /// <summary>
     ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -19,8 +19,19 @@ public class IdValueGenerator : ValueGenerator
     ///     any release. You should only use it directly in your code with extreme caution and knowing that
     ///     doing so can result in application failures when updating to a new Entity Framework Core release.
     /// </summary>
-    public override bool GeneratesTemporaryValues
-        => false;
+    /// <remarks>
+    ///     Do not call this constructor directly from either provider or application code as it may change
+    ///     as new dependencies are added. Instead, use this type in your constructor so that an instance
+    ///     will be created and injected automatically by the dependency injection container. To create
+    ///     an instance with some dependent services replaced, first resolve the object from the dependency
+    ///     injection container, then replace selected services using the C# 'with' operator. Do not call
+    ///     the constructor at any point in this process.
+    /// </remarks>
+    [EntityFrameworkInternal]
+    public CosmosModelRuntimeInitializerDependencies(IJsonIdDefinitionFactory jsonIdDefinitionFactory)
+    {
+        JsonIdDefinitionFactory = jsonIdDefinitionFactory;
+    }
 
     /// <summary>
     ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -28,15 +39,5 @@ public class IdValueGenerator : ValueGenerator
     ///     any release. You should only use it directly in your code with extreme caution and knowing that
     ///     doing so can result in application failures when updating to a new Entity Framework Core release.
     /// </summary>
-    public override bool GeneratesStableValues
-        => true;
-
-    /// <summary>
-    ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-    ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-    ///     any release. You should only use it directly in your code with extreme caution and knowing that
-    ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-    /// </summary>
-    protected override object NextValue(EntityEntry entry)
-        => entry.Metadata.GetJsonIdDefinition()!.GenerateIdString(entry);
+    public IJsonIdDefinitionFactory JsonIdDefinitionFactory { get; }
 }
