+++ src/EFCore.Cosmos/Metadata/Internal/CosmosAnnotationNames.cs
@@ -82,4 +82,12 @@ public static class CosmosAnnotationNames
     ///     doing so can result in application failures when updating to a new Entity Framework Core release.
     /// </summary>
     public const string JsonIdDefinition = Prefix + "JsonIdDefinition";
+
+    /// <summary>
+    ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
+    ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
+    ///     any release. You should only use it directly in your code with extreme caution and knowing that
+    ///     doing so can result in application failures when updating to a new Entity Framework Core release.
+    /// </summary>
+    public const string ModelDependencies = Prefix + "ModelDependencies";
 }
