+++ src/EFCore.Cosmos/Metadata/Internal/IJsonIdDefinition.cs
@@ -1,9 +1,7 @@
-// Licensed to the .NET Foundation under one or more agreements.
+﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using Microsoft.EntityFrameworkCore.Cosmos.Metadata.Internal;
-
-namespace Microsoft.EntityFrameworkCore.Cosmos.ValueGeneration.Internal;
+namespace Microsoft.EntityFrameworkCore.Cosmos.Metadata.Internal;
 
 /// <summary>
 ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -11,7 +9,7 @@ namespace Microsoft.EntityFrameworkCore.Cosmos.ValueGeneration.Internal;
 ///     any release. You should only use it directly in your code with extreme caution and knowing that
 ///     doing so can result in application failures when updating to a new Entity Framework Core release.
 /// </summary>
-public class IdValueGenerator : ValueGenerator
+public interface IJsonIdDefinition
 {
     /// <summary>
     ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -19,8 +17,7 @@ public class IdValueGenerator : ValueGenerator
     ///     any release. You should only use it directly in your code with extreme caution and knowing that
     ///     doing so can result in application failures when updating to a new Entity Framework Core release.
     /// </summary>
-    public override bool GeneratesTemporaryValues
-        => false;
+    IReadOnlyList<IProperty> Properties { get; }
 
     /// <summary>
     ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -28,8 +25,7 @@ public class IdValueGenerator : ValueGenerator
     ///     any release. You should only use it directly in your code with extreme caution and knowing that
     ///     doing so can result in application failures when updating to a new Entity Framework Core release.
     /// </summary>
-    public override bool GeneratesStableValues
-        => true;
+    string GenerateIdString(EntityEntry entry);
 
     /// <summary>
     ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -37,6 +33,5 @@ public class IdValueGenerator : ValueGenerator
     ///     any release. You should only use it directly in your code with extreme caution and knowing that
     ///     doing so can result in application failures when updating to a new Entity Framework Core release.
     /// </summary>
-    protected override object NextValue(EntityEntry entry)
-        => entry.Metadata.GetJsonIdDefinition()!.GenerateIdString(entry);
+    string GenerateIdString(IEnumerable<object?> values);
 }
